Adds tests and helpers for salmon reception.
Fixes one bug where a 200 was not being used as a status code for good
salmon requests.

Three tests.

1) Test that salmon notifications fail when an xrd is not found.

We stub out the host meta such that the webfinger lookup fails to find
anything. We get a resource not found and the code should blow up
gracefully to give us a 404.

2) Test that salmon notifications fail when user verification fails due
to message signing.

Here, we sign a salmon notification with a different user than the one
we check against. Therefore the notification does not match the
signature and thus fails.

3) Test that salmon notifications succeed when verification succeeds.

Now, we sign a salmon notification with the same user we check against.
This should succeed and give us a 200.

Note: I had to fix an old sinatra remnant to get it to work. Begads!
Tests are Good.

Helpers are used to generate and stub out webfinger calls and present
the correct user information for the salmon signing and verification.

Helpers are also used the craft a salmon notification through the
ostatus gem.