Fixes and adds a salmon unit test.
These should be stubs. How did this ever pass?!?

Anyway, added one to check that an empty feed is created for a newly
discovered Author. This is necessary as you follow Feeds, not Authors.
(Since a Feed may not have a discoverable Author, or may have multiple
authors, etc etc etc)