Ensures that remote Feeds are populated even if they've been discovered.
That is, we can discover Feeds in the wild due to salmon notifications.
They are empty. We need a key that indicates they are not subscribed.
That is, they are NOT receiving updates from a hub.

Then, when we do subscribe, we only populate them if there was no prior
relationship with a hub. This solves the problem where Feeds are
discovered through Salmon, but not followed proper.