Adds propeller support.
Propeller is a gem that organizes user defined settings and allows users
to easily define how they want a web application to work.

config/blade.yml - A set of options and addon-modules that the user can
                   select and change.

Addons are pulled from git or rubygems by name using Bundler. Tests are
executed by including the tests from the addons via the minitask rake
test.

Helpers are included in the ApplicationController that can query the
settings of the options.

addon_enabled? :addon_name # Yields true when addon is toggled
user_option_for :option_name # Yields user value of the option, or default

I will remove elasticsearch functionality and place this in an addon
package.