Updates name of edit profile action.
The old name made its destructive behavior unclear and its mutation.
This now has the common update_ prefix and a bang to indicate this. It
is far more instructive than edit_user_profile since edit is an action
and what is that?! It's update everywhere else! Who wrote this, an
alien? :D :D :D